Navbar update and Code Organised
