+++ cpp-website/src/components/Contributors/Contributors.js
@@ -1,5 +1,5 @@
 import React, {useState, useEffect} from "react";
-import "../App.css"
+// import "../App.css"
 // import { Link } from "react-router-dom";
