+++ cpp-website/src/components/Navbar/Navbar.css
@@ -1,0 +1,18 @@
+nav {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #242F9B;
+    height: 85px;
+    width: 100vw;
+    font-size: 1.34rem;
+}
+
+nav div {
+    color: #fff;
+    display: flex;
+    align-items: center;
+    text-decoration: none;
+    padding: 0 2rem;
+    cursor: pointer;
+}
